refactor: aplica principios SOLID e reorganiza testes
