--- .config/dotnet-tools.json
@@ -1,13 +1,0 @@
-{
-  "version": 1,
-  "isRoot": true,
-  "tools": {
-    "csharpier": {
-      "version": "1.2.4",
-      "commands": [
-        "csharpier"
-      ],
-      "rollForward": false
-    }
-  }
-}
